Slightly change the code
The proper input parameter and output control is added in a new
mainvmc2.cpp.

Now the code input various parameters from a parameter file
parameters.txt. For the moment, the code loops through a range of
density and t_tilde so that all data can be got in a single run for a
certain jp. Of course, the ranges can be designed such that each run
only calculate for a certain density and t_tilde.

The way the data is collected is also slightly modified.